mcp-server/pyproject.toml
[project]
name = "mcpress"
version = "0.1.0"
description = "MCP server for MCPress - journalist-written news accessible to AI agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=3.0.0b1",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcpress"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
